fix(multiplexer): change active_session_proxies from usize to u32
Typeshare doesn't support usize type. Changed to u32 which is more
than sufficient for counting session proxies and is TypeScript-compatible.

Changes:
- SystemStatus.active_session_proxies: usize -> u32
- Cast usize result from active_session_proxy_count() to u32

+++ packages/multiplexer/src/api/protocol.rs
@@ -244,7 +244,7 @@ pub struct SystemStatus {
     pub proxies: Vec<ProxyStatus>,
 
     /// Total number of active sessions with proxies
-    pub active_session_proxies: usize,
+    pub active_session_proxies: u32,
 }
 
 /// Request to update a credential
